# speed up iteration over ts